test(typecheck): split Misskey shadow fixtures
