+++ crates/vize/src/commands/check/imports.rs
@@ -338,6 +338,9 @@ mod package_cache_tests;
 #[path = "imports_package_registration_cache_tests.rs"]
 mod package_registration_cache_tests;
 #[cfg(test)]
+#[path = "imports_package_registration_shadow_tests.rs"]
+mod package_registration_shadow_tests;
+#[cfg(test)]
 #[path = "imports_package_registration_tests.rs"]
 mod package_registration_tests;
 #[cfg(test)]
