+++ crates/vize_canon/src/batch/type_checker/tests/generic_props/inline_callback_props.rs
@@ -230,57 +230,6 @@ const jobs = [{ id: 1, timestamp: 2 }]
     );
 }
 
-#[test]
-fn non_generic_child_callback_prop_and_event_inference_stay_contextual() {
-    if resolve_test_tsgo_binary().is_none() {
-        return;
-    }
-    let project_root = create_project_case(
-        "non-generic-inline-callback-prop-event",
-        &[
-            (
-                "src/Child.vue",
-                r#"<script setup lang="ts">
-defineProps<{
-  fileId: string
-  validate?: (file: { id: string }) => boolean | Promise<boolean>
-}>()
-defineEmits<{
-  update: [file: { id: string }]
-}>()
-</script>
-
-<template><span /></template>
-"#,
-            ),
-            (
-                "src/Parent.vue",
-                r#"<script setup lang="ts">
-import Child from './Child.vue'
-let selected = ''
-</script>
-
-<template>
-  <Child
-    file-id="avatar"
-    :validate="async f => f.id.length > 0"
-    @update="f => selected = f.id"
-  />
-</template>
-"#,
-            ),
-        ],
-    );
-
-    let snapshot = snapshot_project_diagnostics(&project_root);
-    let _ = std::fs::remove_dir_all(&project_root);
-    let Some(snapshot) = snapshot else {
-        return;
-    };
-
-    assert_eq!(snapshot, Vec::new());
-}
-
 /// vue-tsc 3.3.4 with TypeScript 6.0.3 reports the three callback-body errors
 /// below verbatim. This covers the two scopes where the resolved prop type must
 /// remain local and a generic inferred only through its callback constraint.
